pyproject.toml
[build-system]
requires = ["setuptools>=77.0.0", "setuptools_scm[toml]>=6.4.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dissect.evidence"
description = "A Dissect module implementing a parsers for various forensic evidence file containers, currently: AD1, ASDF and EWF"
readme = "README.md"
requires-python = ">=3.10"
license = "AGPL-3.0-or-later"
license-files = ["LICENSE", "COPYRIGHT"]
authors = [
  {name = "Dissect Team", email = "dissect@fox-it.com"}
]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "Intended Audience :: Information Technology",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Topic :: Internet :: Log Analysis",
  "Topic :: Scientific/Engineering :: Information Analysis",
  "Topic :: Security",
  "Topic :: Utilities",
]
dependencies = [
    "dissect.cstruct>=4,<5",
    "dissect.util>=3,<4",
]
dynamic = ["version"]

[project.urls]
homepage = "https://dissect.tools"
documentation = "https://docs.dissect.tools/en/latest/projects/dissect.evidence"
repository = "https://github.com/fox-it/dissect.evidence"

[project.optional-dependencies]
full = [
    "pycryptodome",
]
dev = [
    "dissect.cstruct>=4.0.dev,<5.0.dev",
    "dissect.util>=3.0.dev,<4.0.dev",
]

[dependency-groups]
test = [
    "pytest",
    "pycryptodome",
]
lint = [
    "ruff==0.13.1",
    "vermin",
]
build = [
    "build",
]
debug = [
    "ipdb",
]
dev = [
    {include-group = "test"},
    {include-group = "lint"},
    {include-group = "debug"},
]

[project.scripts]
asdf-dd = "dissect.evidence.tools.asdf.dd:main"
asdf-meta = "dissect.evidence.tools.asdf.meta:main"
asdf-repair = "dissect.evidence.tools.asdf.repair:main"
asdf-verify = "dissect.evidence.tools.asdf.verify:main"

adcrypt = "dissect.evidence.tools.adcrypt:main"

[tool.ruff]
line-length = 120
required-version = ">=0.13.1"

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = [
  "F",
  "E",
  "W",
  "I",
  "UP",
  "YTT",
  "ANN",
  "B",
  "C4",
  "DTZ",
  "T10",
  "FA",
  "ISC",
  "G",
  "INP",
  "PIE",
  "PYI",
  "PT",
  "Q",
  "RSE",
  "RET",
  "SLOT",
  "SIM",
  "TID",
  "TCH",
  "PTH",
  "PLC",
  "TRY",
  "FLY",
  "PERF",
  "FURB",
  "RUF",
]
ignore = ["E203", "B904", "UP024", "ANN002", "ANN003", "ANN204", "ANN401", "SIM105", "TRY003"]

[tool.ruff.lint.per-file-ignores]
"tests/_docs/**" = ["INP001"]
"*.pyi" = ["E", "F", "PYI"]

[tool.ruff.lint.isort]
known-first-party = ["dissect.evidence"]
known-third-party = ["dissect"]

[tool.setuptools.packages.find]
include = ["dissect.*"]

[tool.setuptools_scm]
